Update to version 2.0
This plugin is designed to optimize images upon upload and create webp versions, enhancing the website's performance. It includes functions to compress JPEG and PNG images using tools like jpegoptim, optipng, and pngquant. The plugin also creates webp versions of the images and manages them, including deleting orphaned webp files. A daily scheduled event checks and manages webp images if a specific constant 'VWPIMGOP' is defined as true in the WordPress configuration. Additionally, the plugin ensures that webp versions are deleted when the original images are removed from the media library.